requirements.txt
click==8.1.3
distlib==0.3.6
filelock==3.9.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
platformdirs==2.6.2
python-dotenv==0.21.1
virtualenv==20.17.1
Werkzeug==2.2.2
